requirements.txt
click==8.1.8
colorama==0.4.6
mysql-connector==2.2.9
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
pyqt6-tools==6.4.2.3.3
PyQt6_sip==13.9.1
python-dotenv==1.0.1
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
